+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [`12.0.1.dev0` (unreleased)](https://github.com/kdeldycke/extra-platforms/compare/v12.0.0...main)
+
+> [!WARNING]
+> This version is **not released yet** and is under active development.
+
 ## [`12.0.0` (2026-04-24)](https://github.com/kdeldycke/extra-platforms/compare/v11.1.0...v12.0.0)
 
 - Add `SH` (Bourne Shell) trait and `is_sh()` detection function.